LibGraphics: Implemented TexturedEntity
Added another entity extending the Entity class that is optimized
for handling textures.
Also moved all entity related code into the Entity directory.
As TexturedEntity uses a texture and does not use a color like the
base Entity class a 'default' color (white) was defined in Common
as this might also come in handy in other places.