Switched to conan and implemented random level generation
As it seems tedious to always initialize vcpkg, I switched the whole
thing over to conan which seems to be more easy to initialize and
maintain.

Also implemented random level generation based on the current window
dimensions.

There is still a bug with the brick collisions as the ball is sometimes
just going through all of them, also the paddle is kinda hard to
control.